Update README.md
This is a number guessing program. This firstly asks for the difficulty from the user, and according to the difficulty, it will provide a number of attempts to guess the number. In case the user fails to guess the number, "GAME OVER" will be displayed. To make it more interactive, I also added colors.